LayoutTests/platform/ios-simulator/editing/pasteboard/smart-paste-002-expected.txt

EDITING DELEGATE: shouldBeginEditingInDOMRange:range from 0 of DIV > BODY > HTML > #document to 3 of DIV > BODY > HTML > #document
EDITING DELEGATE: webViewDidBeginEditing:WebViewDidBeginEditingNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: shouldChangeSelectedDOMRange:range from 0 of DIV > DIV > BODY > HTML > #document to 0 of DIV > DIV > BODY > HTML > #document toDOMRange:range from 5 of #text > DIV > DIV > BODY > HTML > #document to 5 of #text > DIV > DIV > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: shouldInsertNode:#document-fragment replacingDOMRange:range from 0 of DIV > DIV > BODY > HTML > #document to 0 of DIV > DIV > BODY > HTML > #document givenAction:WebViewInsertActionPasted
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: shouldChangeSelectedDOMRange:range from 6 of #text > SPAN > DIV > DIV > BODY > HTML > #document to 6 of #text > SPAN > DIV > DIV > BODY > HTML > #document toDOMRange:range from 0 of #text > FONT > DIV > BODY > HTML > #document to 6 of #text > SPAN > DIV > DIV > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
EDITING DELEGATE: shouldApplyStyle:<DOMCSSStyleDeclaration: 0x7fc61c89b030> toElementsInDOMRange:range from 0 of #text > FONT > DIV > BODY > HTML > #document to 6 of #text > SPAN > DIV > DIV > BODY > HTML > #document
EDITING DELEGATE: shouldChangeSelectedDOMRange:range from 2 of DIV > BODY > HTML > #document to 6 of #text > SPAN > FONT > DIV > DIV > BODY > HTML > #document toDOMRange:range from 0 of #text > SPAN > DIV > BODY > HTML > #document to 6 of #text > SPAN > FONT > DIV > DIV > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
Tests: 
Smart paste when pasting at the beginning of a line before a word.
Expected Results: 
A space should be added after the pasted word before the preexising word. It should like this: 
test test
hello 
world
WebKittest
execCopyCommand: <div id="test" class="editing"> test </div>
execPasteCommand: <span style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);">hello&nbsp;<br></span><div style="-webkit-line-break: after-white-space;"><div><span style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);">world</span></div></div><div id="test" class="editing"><font size="3"><span style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);">WebKit</span></font>test </div>
